requirements.txt
google-cloud-aiplatform==1.48.0
streamlit==1.33.0
PyYAML==6.0.1
fake_useragent
langchain==0.1.16
pypdf==4.2.0
chromadb==0.5.0
youtube-transcript-api==0.6.2
langchain-community==0.0.34
pytube==15.0.0
